Fix topsort dropping mapping function definitions
Splicing in alternative definitions uses a topsort, but early enough that
the mappings have not been rewritten away.  Manually written mapping
functions were previously lost because the callgraph didn't have back
edges from the function to the mapping, so the functions weren't in the
same component as the val_spec, but the sort would assume that
including the val_spec component was enough to define the functions.